requirements.txt
aniso8601==8.0.0
astroid==2.4.2
attrs==19.3.0
Babel==2.8.0
blinker==1.4
click==7.1.2
colorama==0.4.3
dnspython==2.0.0
email-validator==1.1.1

Flask-BabelEx==0.9.4

Flask-Mail==0.9.1

Flask-Principal==0.4.0
flask-restplus==0.13.0
Flask-Security==3.0.0
Flask-WTF==0.14.3
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1

passlib==1.7.2
pylint==2.5.3

pyrsistent==0.16.0
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
speaklater==1.3
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
